fix build

+++ src/tests/test_core_operators.cc
@@ -25,7 +25,7 @@ TEST(Operator, ReadOperator) {
   reader.start(config);
   boost::this_thread::sleep(boost::posix_time::seconds(1));
 
-  ASSERT_GT(rec->tuples.size(), 4);
+  ASSERT_GT(rec->tuples.size(),(unsigned int) 4);
   string s = rec->tuples[0].e(0).s_val();
   ASSERT_TRUE(s.length() > 0 && s.length() < 100); //check that output is a sane string
   ASSERT_NE(s[s.length() -1], '\n'); //check that we prune \n.
